fix: 스토리 공유 완료 alert창 깨지는 현상 해결
- 부모 요소에 absolute 속성이 있어, 그 하위에서 fixed 속성을 부여해도 부모를 기준으로 보여지는 문제가 있었습니다.
- fixed 속성을 가지는 CustomAlert 컴포넌트를 상위 컴포넌트(기존 부모 컴포넌트)로 옮겼습니다.
