+++ src/app/story/[id]/_components/controller/StoryControlBar.tsx
@@ -1,6 +1,8 @@
 import { Dispatch, SetStateAction } from "react";
 import MenuBar from "./MenuBar";
 import NavigateBtn from "./NavigateBtn";
+import CustomAlert from "@/components/ui/CustomAlert";
+import { useCustomAlert } from "@/hooks/useCustomAlert";
 
 type StoryControlBarProps = {
   liked: boolean;
@@ -27,8 +29,37 @@ const StoryControlBar = ({
   onNext,
   onPrev,
 }: StoryControlBarProps) => {
+  const { isOpen, options, openAlert, closeAlert } = useCustomAlert();
+
+  const handleLinkShare = async () => {
+    try {
+      await navigator.clipboard.writeText(window.location.href);
+      openAlert({
+        title: "링크 복사 완료",
+        message: "링크가 복사되었습니다!",
+        type: "success",
+      });
+    } catch (err) {
+      console.error("링크 복사 실패", err);
+      openAlert({
+        title: "링크 복사 실패",
+        message: "링크 복사에 실패했습니다",
+        type: "error",
+      });
+    }
+  };
+
   return (
     <>
+      <CustomAlert
+        isOpen={isOpen}
+        title={options.title}
+        message={options.message}
+        type={options.type}
+        onConfirm={options.onConfirm || closeAlert}
+        onCancel={options.onCancel}
+      />
+
       {/* 메뉴바 */}
       <div className="absolute left-4 bottom-20 bg-white bg-opacity-95 rounded-2xl shadow-lg backdrop-blur-sm">
         <MenuBar
@@ -40,9 +71,9 @@ const StoryControlBar = ({
           setClothesOn={setClothesOn}
           postComment={postComment}
           setPostComment={setPostComment}
+          onShare={handleLinkShare}
         />
       </div>
-
       {/* 스토리 전환 버튼 */}
       <div className="space-y-10 absolute top-2/5 right-4">
         <NavigateBtn onNextPage={onNext} onPrevPage={onPrev} />
